mooooore minor fixes

+++ osrefl/theory/DWBA_Cuda.py
@@ -244,9 +244,9 @@ def scatCalc(cell,lattice,beam,q):
             # Call DWBA function on the GPU
             cudaDWBA(q.q_list[0][i], q.q_list[1][ii],
                      numpy.int32(cell.step[0]), numpy.int32(cell.step[1]),
-                      numpy.int32(size(x[0])),size(y[0]),
+                     numpy.int32(size(x[0])), numpy.int32(size(y[0])),
                      cxx, cyy, crtor,
-                     Vfac, coutput,
+                     numpy.float32(Vfac), coutput,
                      block=(400,1,1), grid=(1,1))
